Create the sequence headers renaming files and update Compi
This commit adds a slight modification of the rename headers
script to produce also renaming files containing the mappings
from the original sequence headers. This allows users of the
pipeline to know exactly the names of the sequences in
intermediate renamed files that can be produced by the
pipeline.

This commit also updates Compi to 1.3.4.