Fix the prepare tres for codeML step
This commit fixes the prepare trees for codeML step to avoid
issues with concurrent seds on the same directory. This is
because it was noticed that using a high number of concurrent
Compi tasks (e.g. 50) caused some trees to not be created
without any obvious reason. However, they were created when
running only the foreach tasks for only those files.